deps: update V8 to 6.1.534.36
PR-URL: https://github.com/nodejs/node/pull/14730

+++ deps/v8/test/unittests/base/platform/platform-unittest.cc
@@ -154,7 +154,8 @@ class MemoryAllocationPermissionsTest : public ::testing::Test {
     if (!sigsetjmp(continuation_, save_sigs)) {
       switch (action) {
         case MemoryAction::kRead: {
-          USE(*buffer);
+          // static_cast to remove the reference and force a memory read.
+          USE(static_cast<int>(*buffer));
           break;
         }
         case MemoryAction::kWrite: {
